Remove explicit handles in StartSessionAuth
This commit adds a few fixes. The main improvement is due to a bugfix
being published upstream in a release. This bugfix allows auth sessions
to be started with the help of encryption sessions (so that parameters
are encrypted within the command). Thus, StartAuthSession no longer
needs to receive separate sessions handles and can now simply use the
ones stored on the Context. This does, however, mean that the most
recent version of the TSS libraries is needed.
Another fix is for memory allignment for 32bit machines for
TPMS_CONTEXT. When generating the bindings on 32 bit architectures, a
padding field that did not exist on 64 bit machines is added. Any
attempt to use the One True Constructor for said structure then fails if
the new field is not considered.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>