Added more TPM resource API:s
Added more of the ESYS TOM Resource API:s. Some of the API:s already existed with
some extended functionality or a different name. I changed this to call new functions
that are only thin wrappers around the original C API:s. The reason for the changes is
to keep the naming consistent.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>