Improved TpmaSession
The TpmaSession have now been split up into two parts.
The first part the actual TpmaSession struct now only
holds the values for the mask and the flags and that is
all it is responsible for.
The TpmaSessionBuilder struct provides utility
functions that makes it easier to create a TpmaSession
object.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>