Added more TPM Resource API:s (#54)
* Added more TPM resource API:s

Added more of the ESYS TOM Resource API:s. Some of the API:s already existed with
some extended functionality or a different name. I changed this to call new functions
that are only thin wrappers around the original C API:s. The reason for the changes is
to keep the naming consistent.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>

* Fixed names.

Fixed the names of the TPM2 Resource API wrapper functions.
Removed the duplicates previously created in order to be
backward compatible.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>

* Fixed name change in tests.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>

* Improved TpmaSession

The TpmaSession have now been split up into two parts.
The first part the actual TpmaSession struct now only
holds the values for the mask and the flags and that is
all it is responsible for.
The TpmaSessionBuilder struct provides utility
functions that makes it easier to create a TpmaSession
object.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>

* Fixed tests that failed to build

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>