Fixed interface types used for nv context methods.
The nv_define_space and nv_undefine_space were using
the NvAuth interface type. Though according to the specification
they should be using the Provision interface type.

The other nv context methods were simply using an auth handle
that could be anything. This is probably not correct. According to
the TPM specification they should be using the NvAuth interface type.

These issues were mentioned in #184 and should with this be resolved.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>