This is an attempt to fix tests that fails localy.
Sometimes when trying to run tests localy they fail with
som cryptic message that a new auth session could not be
created due to authorization failure.

The cause behind this is that one session is being used
as an optional session for creating another session
wether this should be allowed or not is not quite clear.
Until this have been fully investigated this change remove
this behaviour.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>