Abstract execute_With_nullauth_session return type
This makes it possible to call any function from
execute_with_nullauth_session, as long as the resulting Error type can
be created from tss_esapi::Error.
This allows an application to call a function that returns myapp::Error,
as long as an implementation From<tss_esapi::Error> exists for that
type.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>