Ensure PcrData retains order
PcrData would (possibly) reorder and merge the different TPM2B_DIGEST
entries in a TPML_DIGEST.
This is not valid in the case where for example the TPML_DIGEST is the
result of a pcr_read call: in that case, the order of the
TPML_PCR_SELECTION entries and the TPML_DIGEST match, and must remain
matching.
This is important when a digest of all the digests is passed along, like
in the case of a Quote operation.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>