Move TransientKeyContext away from contexts
This commit moves the TransientKeyContext away from using saved key
contexts. Instead, keys are returned and managed as pairs of public and
(encrypted) private parts which are loaded into the TPM every time an
operation is requested.
A migration mechanism is also implemented to convert from contexts to
the new key material format.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>