Fix compile error when using zeroize_derive >= 1.2.0
Fix "cannot move out of type which implements the Drop trait" compile
errors when using zeroize_derive >= 1.2.0 by matching on a reference
to the signature field contained in the SignatureData type. The code
previously compiled fine when using zeroize_derive <= 1.1.0 as it was
mistakenly not deriving the Drop trait for enum types.

Fixes #260

Signed-off-by: Rob Shearman <rob@graphiant.com>