Fix handle leak in abstraction::nv::list
The call to tr_from_tpm_public public in an earlier map operation in
list returns a handle that should be closed when no longer required,
so to avoid a leak close this whilst iterating over the results.

To make the function easier to read, collapse the multiple map
operations into one and extract its functionality out into a separate
function that returns the NvPublic and Name associated with an NV
index TPM handle.

Signed-off-by: Rob Shearman <rob@graphiant.com>