Add a hint for unique_identifier functions
Previously a unique identifier was not explicitly required and if not
provided a default value was used. After migration it is not clear what
value need to be passed so this patch improves the error message
suggesting the default values as that will preserve old behavior in
a code that tries to upgrade to tss-esapi 7.x.

Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>