Fix usize conversion
This commit fixes the conversion from `TPMT_PUBLIC` size to the input of
Tss2_MU_TPMT_PUBLIC_Marshal. A new WrapperErrorKind variant is added for
internal errors which were not expected (essentially as a replacement
for `unwrap`).

The cross-compilation script was also moved to run per-commit in the CI
instead of nightly - the fix above would've been obvious if we tried
cross-compilation.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>