Allow a Public object to be marshalled/unmarshalled
When performing remote attestation then it is often necessary to
transmit certain TPM objects over a network to the remote entity and
the Public object is one of them, so being able to marshall/unmarshall
is useful.

Since the Public type more represents TPMT_PUBLIC than TPM2B_PUBLIC
then change the comment for the type and add conversions to/from
TPMT_PUBLIC. Then implement the Marshall and UnMarshall traits for the
TPMT_PUBLIC representation.

Add tests for marshalling/unmarshalling public data.

Signed-off-by: Rob Shearman <rob@graphiant.com>