Allow a Signature to be marshalled/unmarshalled
When performing remote attestation then it is often necessary to
transmit certain TPM objects over a network to the remote entity and
the Signature object is one of them, so being able to convert this
to/from a bytes representation is useful.

So in order to support sending Signature objects to a remote entity
over a network, implement the Marshall and UnMarshall traits.

Signed-off-by: Rob Shearman <rob@graphiant.com>