Add Read trait implementation for nv abstraction
There is an existing read_full function that abstracts non-volatile
storage reading, but it is specific to the one task and is hard to use
more generically.

So implement std::io::Read trait for non-volatile storage by adding an
NvReaderWriter object which the trait is implemented for, allowing the
object to be obtained through an NvOpenOptions builder object. Using
the std::io::Read trait exposes lower-level functionality meaning it is
more flexible, but the trait also provides higher-level functionality
as well, with the only downside being potentially it not being as
efficient. It also allows the use of TPM non-volatile storage in any
place that a Read trait is implemented, such as std::io::copy.

Implement the existing read_full function in terms of the
NvOpenOptions builder, NvReaderWriter::size and
NvReaderWriter::read_to_end, which is provided by the Read trait.

Signed-off-by: Rob Shearman <rob@graphiant.com>