Add Seek and Write trait implementations for nv abstraction
Implement std::io::Write trait for non-volatile storage on the
NvReaderWriter object. Using the std::io:Write trait exposes
lower-level functionality versus a stand-alone function meaning it is
more flexible, but the trait also provides higher-level functionality
as well.

Given that non-volatile storage is random access, implement
std::io:Seek as well, giving the benefit of being able to seek to
arbitrary points, as well as query the size of the storage entry.

Signed-off-by: Rob Shearman <rob@graphiant.com>