Fixes PcrSelectSize parse bug.
Fixes a bug when parsing TPM 'size of select' where it previously
would allow for a value that is larger then the maximum size of the
buffer that should hold the octets and this would hav resulted in a
runtime error.

Hides the num_derives dependency by adding try_parse- and as- methods
for a number of types.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>