Improves 'use' and naming in context methods.
Removes all 'use crate::tss2_esys::*' and replaces it with more
specific use declarations that only imports with module needs.

Makes the naming of variables in the context methods more consistent by
allways using the name that are snake case versions of the names used
in the declaration of the Tss_Esys-functions.
A '_ptr' suffix is used when it is a variable that will contain
a pointer to a memory that is not freed by the tpm2-tss library.

Makes use of the YesNo interface for handling the conversions from bool
to TPMI_YES_NO.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>