Removes code duplication in named_field_buffer_type macro.
Adds an associated function to the buffer types that are created by
the named_field_buffer_type macro. This is then used in the TryFrom
implementations to verify the size. Previously very similar code was
used in the different TryFrom implementations this has now been replaced
with usage of the added associated function.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>