Add unmarshall_offset to UnMarshall trait
Adding a second method to the UnMarshall trait - unmarshall_offset -
which takes a second parameter, a mutable offset, which is modified
during unmarshalling to point to the first position in the input buffer
that was not used for unmarshalling. The new method can be easily used
to implement the basic `unmarshall` method by passing `&mut 0` as the
offset.

Also adding `Marshall` and `UnMarshall` for `CommandCode` and `u32`.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>