Bump MSRV and fix clippy lints
Bumping MSRV to 1.60.0 as required for the picky-asn1 crate.

Also setting the MSRV for clippy and fixing some of the remaining
warnings. Some of the things clippy complains about for newer versions
of the compiler can be fixed/changed without losing support for older
compiler versions, but some features are incompatible. Since clippy is
now configured to understand the MSRV, there's also no need for some of
the `allow` attributes.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>