Fix MSRV-related issues
Fixing the MSRV in our CI involves making sure that the crate can build
under that version of the Rust compiler, given some choice of
dependencies that are also compatible with the MSRV. A major issue is
that many crates bump the MSRV in patch or minor releases that then get
picked up automatically if no lock file is present.

This commit adds a lock file for our current MSRV, and pins the MSRV for
clippy as well.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>