Add definition for UndefineSpaceSpecial
Add a implementation for the UndefineSpaceSpecial function which is used
to delete an NV index that has the TPMA_NV_POLICY_DELETE attribute set.

An example and test are included for this function, setting up the NV
index and deleting it requires additional initialization of a policy
digest as well as a policy session in order to perform the delete due to
the policy requiring the NvUndefineSpaceSpecial command code defined.

Signed-off-by: Nathan Rossi <nathan.rossi@digi.com>