Disables the clippy multiple_crate_versions lint error.
In Rust version 1.77 clippy has begun reporting
errors for when two versions of the same crate
exists among the dependencies. In this case
it is the `syn` crate for which there exists two
versions. This is to the reason that the
`zeroize_derive` crate updated the `syn` version
after making a Rust Edition bump. So the ones
using Rust Edition 2018 gets an older version
of `eroize_derive` which uses `syn` with major
release version 1.

Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>