Adds a TBS TctiNameConf.
When #477 got merged it became possible to build
using a path to the ```tpm2-tss``` installation
instead of depending on ```pkg-config```.

This made it possible to build under Windows. To
further increase the support for the windows
platform this commit moves the option for TBS
TCTI that is being introduced in #523 into a
separate commit.

This commit also updates the documentation
regarding building using an installation
folder.

Co-authored-by: Thomas Epperson <thomas.epperson@gmail.com>
Signed-off-by: Jesper Brynolf <jesper.brynolf@gmail.com>