Add definition for NV_Extend
Add a implementation for NV_Extend, which is used with a NV Index
defined with the NvIndexType::Extend to provide a PCR equivalent that
uses NV memory. The hashing algorithm used is defined by the index name
algorithm.

An example and test are included for this function, setting up the NV
index with the required type and typical attributes and performing an
extend of the index and confirming the resulting hash value.

Signed-off-by: Nathan Rossi <nathan.rossi@digi.com>