Add definition for NV_Extend
Add a implementation for NV_Extend, which is used with a NV Index
defined with the NvIndexType::Extend to provide a PCR equivalent that
uses NV memory. The hashing algorithm used is defined by the index name
algorithm.

An example is included within the documentation for the function,
setting up the NV index with the required type and typical attributes
for performing a extend of the index. An integration test is also added
for this function. Testing the same owner auth scenario of the example,
and a platform auth variant, validating the result of the NV index after
the extend operation in both tests.

Signed-off-by: Nathan Rossi <nathan.rossi@digi.com>